inherit OutputMapper from OutputMapperBase
