+++ algorithms/base/__init__.py
@@ -1,0 +1,1 @@
+from .output_mapper import OutputMapperBase
