+++ algorithms/base/input_mapper.py
@@ -1,0 +1,9 @@
+class InputMapperBase:
+    def __init__(self,):
+        pass
+
+    def format_sequence(sequence):
+        return sequence
+
+    def format_input(spectrum): # file_i? 
+        return spectrum
