+++ token_masses.py
@@ -5,6 +5,7 @@ AA_MASSES = {
     "P": 97.052763875,
     "V": 99.068413945,
     "T": 101.047678505,
+    "C": 103.009184505,
     "C+57.021": 160.030644505,
     "L": 113.084064015,
     "I": 113.084064015,
